Use the encapsulated task refactor of joining-tasks example
