--- tests/dummy/app/docs/examples/joining-tasks-2/template.hbs
@@ -1,31 +1,0 @@
-<h3>Awaiting Multiple Child Tasks (or, cancelable Promise.all, Promise.race)</h3>
-
-<p>
-  This is a rewrite of the <LinkTo @route="docs.examples.joining-tasks">Awaiting Multiple Child Tasks</LinkTo>
-  example, using encapsulated-tasks and completion state on task objects.
-</p>
-
-<h3>Live Example</h3>
-
-<h5>Status: {{this.status}}</h5>
-
-{{! BEGIN-SNIPPET joining-tasks-2-template }}
-<p>
-  <button onclick={{perform this.parent "all"}} type="button">all()</button>
-  <button onclick={{perform this.parent "race"}} type="button">race()</button>
-</p>
-{{! END-SNIPPET }}
-
-{{#each this.childTasks as |ti|}}
-  <div class="progress-outer">
-    <div class="progress-inner" style={{progress-style ti.percent ti.id this.colors}} {{! template-lint-disable no-inline-styles }}>
-      Progress: {{ti.percent}}%
-      {{#if ti.value}}
-        Word: {{ti.value}}
-      {{/if}}
-    </div>
-  </div>
-{{/each}}
-
-<CodeSnippet @name="joining-tasks-2.js" />
-<CodeSnippet @name="joining-tasks-2-template.hbs" />
